rename & extract class
