+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/preset/PresetEventLocation.scala
@@ -1,0 +1,7 @@
+package x7c1.linen.modern.init.settings.preset
+
+sealed trait PresetEventLocation
+
+case object PresetTabSelected extends PresetEventLocation
+
+case object PresetTabAll extends PresetEventLocation
